requirements.txt
# Google APIs
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Trello
py-trello==0.19.0

# Plaud (unofficial API client)
requests==2.32.3

# Email
sendgrid==6.11.0

# AI briefing generation
anthropic==0.34.0

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.1

# Utilities
pytz==2024.1
jinja2==3.1.4
